Check if path exists before linking.

+++ docker/driver-loader/docker-entrypoint.sh
@@ -50,7 +50,8 @@ echo "* Setting up /usr/src links from host"
 
 for i in "$HOST_ROOT/usr/src"/*
 do
-    base=$(basename "$i")
+    [[ -e $i ]] || continue
+		base=$(basename "$i")
     ln -s "$i" "/usr/src/$base"
 done
